Update sdk_tools.py
if there were multiple 'id' fields with the same last part in the action data ID, the function could potentially replace the token with the wrong 'id' field, or not at all.

For example, consider two 'id' fields: PLUGIN_ID + ".act.light.selection.brightness" and PLUGIN_ID + ".act.light.selection.custom.brightness". Both of these 'id' fields have the same last part: "brightness". If the token was "brightness", the function could replace it with either of these 'id' fields, leading to incorrect results.

Fix involved modifying the function to match the token with the end of the 'id' field. If a match is found, the token is replaced with the corresponding 'id' field. This ensures that the correct 'id' field is used for replacement, even if multiple 'id' fields have the same last part.

User is now able to use multiple words to identify a data action ID as well.. previously you would use $[brightness]  but if they had two in the same data actions it would fail.. so now you can use $[custom.brightness]